Added PHP script
Added a PHP script that computes the move of the computer. The move is random unless one of the two players can win the game. If the computer can win then the move will be to win but if the human player have the opportunity to win then the computer will deny the win by moving in the decisive cell.