Added hard difficulty and reset stats button
Now the game has two difficulties: easy and hard. Each difficulty represents a different request to the PHP server in the back-end of the application. I also added a reset statistics button that resets the number of wins/victories, the number of loses/defeats and the number of draws to 0.